pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "securepremium"
version = "0.1.0"
description = "SecurePremium - Quantify device compromise risk as insurance premiums"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Security Team", email = "security@example.com"}
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
]

dependencies = [
    "device-fingerprinting-pro>=2.2.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "requests>=2.28.0",
    "cryptography>=41.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/example/securepremium"
Repository = "https://github.com/example/securepremium"
Issues = "https://github.com/example/securepremium/issues"
Documentation = "https://github.com/example/securepremium/blob/main/README.md"
